Move locale repository to backend-api module
