--- dc-cudami-admin/dc-cudami-admin-backend-rest/src/main/java/de/digitalcollections/cudami/admin/backend/impl/repository/LocaleRepository.java
@@ -1,20 +1,0 @@
-package de.digitalcollections.cudami.admin.backend.impl.repository;
-
-import feign.RequestLine;
-import java.util.List;
-import java.util.Locale;
-
-public interface LocaleRepository {
-
-  @RequestLine("GET /latest/languages")
-  List<String> findAllLanguages();
-
-  @RequestLine("GET /latest/languages/default")
-  Locale getDefaultLanguage();
-
-  @RequestLine("GET /latest/locales")
-  List<String> findAllLocales();
-
-  @RequestLine("GET /latest/locales/default")
-  String getDefaultLocale();
-}
