--- dc-cudami-admin/dc-cudami-admin-backend-rest/src/main/java/de/digitalcollections/cudami/admin/backend/impl/repository/LocaleRepositoryConfig.java
@@ -1,31 +1,0 @@
-package de.digitalcollections.cudami.admin.backend.impl.repository;
-
-import com.fasterxml.jackson.databind.ObjectMapper;
-import de.digitalcollections.commons.feign.codec.EndpointErrorDecoder;
-import feign.Feign;
-import feign.jackson.JacksonDecoder;
-import feign.jackson.JacksonEncoder;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class LocaleRepositoryConfig {
-
-  @Value(value = "${cudami.server.address}")
-  private String cudamiServerAddress;
-
-  @Autowired
-  ObjectMapper objectMapper;
-
-  @Bean
-  public LocaleRepository localeRepository() {
-    LocaleRepository endpoint = Feign.builder()
-        .decoder(new JacksonDecoder(objectMapper))
-        .encoder(new JacksonEncoder(objectMapper))
-        .errorDecoder(new EndpointErrorDecoder())
-        .target(LocaleRepository.class, cudamiServerAddress);
-    return endpoint;
-  }
-}
